requirements.txt

bidict==0.23.1
certifi==2024.7.4
charset-normalizer==3.3.2
cycler==0.12.1
dnspython==2.6.1
eventlet==0.36.1
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.5.1
networkx==2.8
numpy==1.22.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pyasn1==0.6.0
pyparsing==3.1.2
PySimpleGUI==5.0.6
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
threadpoolctl==3.5.0
torch==1.8.1
tqdm==4.66.5
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wsproto==1.2.0

